+++ sites/partners/pages/listings/[id]/applications/resolved/index.tsx
@@ -1,36 +1,23 @@
-import React, { useContext } from "react"
+import React from "react"
 import { useRouter } from "next/router"
 import Head from "next/head"
 import {
   AgTable,
   t,
-  Button,
-  SiteAlert,
   useAgTable,
   Breadcrumbs,
   BreadcrumbLink,
   NavigationHeader,
 } from "@bloom-housing/ui-components"
-import { AuthContext } from "@bloom-housing/shared-helpers"
-import {
-  useSingleListingData,
-  useFlaggedApplicationsList,
-  useApplicationsExport,
-} from "../../../../../lib/hooks"
+import { useSingleListingData, useFlaggedApplicationsList } from "../../../../../lib/hooks"
 import { ListingStatusBar } from "../../../../../src/listings/ListingStatusBar"
 import Layout from "../../../../../layouts"
 import { ApplicationsSideNav } from "../../../../../src/applications/ApplicationsSideNav"
-import { tableColumns, getLinkCellFormatter } from "../../../../../src/applications/helpers"
+import { getLinkCellFormatter } from "../../../../../src/applications/helpers"
 
 const ApplicationsList = () => {
-  const { profile } = useContext(AuthContext)
   const router = useRouter()
   const listingId = router.query.id as string
-
-  const { onExport, csvExportLoading, csvExportError } = useApplicationsExport(
-    listingId,
-    profile?.roles?.isAdmin ?? false
-  )
 
   const tableOptions = useAgTable()
 
@@ -41,7 +28,55 @@ const ApplicationsList = () => {
     listingId,
     page: 1,
     limit: 1,
+    view: "resolved",
   })
+
+  const columns = [
+    {
+      headerName: t("applications.duplicates.duplicateGroup"),
+      field: "id",
+      sortable: false,
+      filter: false,
+      pinned: "left",
+      cellRenderer: "formatLinkCell",
+      valueGetter: ({ data }) => {
+        if (!data?.applications?.length) return ""
+        const applicant = data.applications[0]?.applicant
+
+        return `${applicant.firstName} ${applicant.lastName}: ${data.rule}`
+      },
+    },
+    {
+      headerName: t("applications.duplicates.primaryApplicant"),
+      field: "",
+      sortable: false,
+      filter: false,
+      pinned: "left",
+      valueGetter: ({ data }) => {
+        if (!data?.applications?.length) return ""
+        const applicant = data.applications[0]?.applicant
+
+        return `${applicant.firstName} ${applicant.lastName}`
+      },
+    },
+    {
+      headerName: t("applications.pendingReview"),
+      field: "",
+      sortable: false,
+      filter: false,
+      pinned: "left",
+      valueGetter: ({ data }) => {
+        return `${data?.applications?.length ?? 0}`
+      },
+    },
+    {
+      headerName: t("t.rule"),
+      field: "rule",
+      sortable: false,
+      filter: false,
+      pinned: "left",
+    },
+  ]
 
   return (
     <Layout>
@@ -57,7 +92,6 @@ const ApplicationsList = () => {
           flagsQty: flaggedAppsData?.meta?.totalFlagged,
           listingLabel: t("t.listingSingle"),
           applicationsLabel: t("nav.applications"),
-          flagsLabel: t("nav.flags"),
         }}
         breadcrumbs={
           <Breadcrumbs>
@@ -71,13 +105,7 @@ const ApplicationsList = () => {
             </BreadcrumbLink>
           </Breadcrumbs>
         }
-      >
-        {csvExportError && (
-          <div className="flex top-4 right-4 absolute z-50 flex-col items-center">
-            <SiteAlert type="alert" timeout={5000} dismissable />
-          </div>
-        )}
-      </NavigationHeader>
+      />
 
       <ListingStatusBar status={listingDto?.status} />
 
@@ -97,14 +125,14 @@ const ApplicationsList = () => {
               }}
               config={{
                 gridComponents: { formatLinkCell: getLinkCellFormatter(router) },
-                columns: tableColumns,
+                columns: columns,
                 totalItemsLabel: t("applications.totalApplications"),
               }}
               data={{
-                items: flaggedAppsData.items,
+                items: flaggedAppsData?.items ?? [],
                 loading: flaggedAppsLoading,
-                totalItems: flaggedAppsData.meta.totalItems,
-                totalPages: flaggedAppsData.meta.totalPages,
+                totalItems: flaggedAppsData?.meta?.totalItems ?? 0,
+                totalPages: flaggedAppsData?.meta?.totalPages ?? 0,
               }}
               search={{
                 setSearch: tableOptions.filter.setFilterValue,
@@ -112,13 +140,6 @@ const ApplicationsList = () => {
               sort={{
                 setSort: tableOptions.sort.setSortOptions,
               }}
-              headerContent={
-                <div className="flex-row">
-                  <Button className="mx-1" onClick={() => onExport()} loading={csvExportLoading}>
-                    {t("t.export")}
-                  </Button>
-                </div>
-              }
             />
           </div>
         </article>